third_party/blink/web_tests/external/wpt/css/css-inline/inheritance-expected.txt

This is a testharness.js-based test.
FAIL Property alignment-baseline has initial value baseline assert_equals: expected "baseline" but got "auto"
PASS Property alignment-baseline does not inherit
PASS Property baseline-shift has initial value 0px
PASS Property baseline-shift does not inherit
FAIL Property dominant-baseline has initial value normal assert_equals: expected "normal" but got "auto"
PASS Property dominant-baseline inherits
FAIL Property initial-letters has initial value normal assert_own_property: expected property "initial-letters" missing
FAIL Property initial-letters does not inherit assert_own_property: expected property "initial-letters" missing
FAIL Property initial-letters-align has initial value alphabetic assert_own_property: expected property "initial-letters-align" missing
FAIL Property initial-letters-align inherits assert_own_property: expected property "initial-letters-align" missing
FAIL Property initial-letters-wrap has initial value none assert_own_property: expected property "initial-letters-wrap" missing
FAIL Property initial-letters-wrap inherits assert_own_property: expected property "initial-letters-wrap" missing
FAIL Property initial-sizing has initial value normal assert_own_property: expected property "initial-sizing" missing
FAIL Property initial-sizing does not inherit assert_own_property: expected property "initial-sizing" missing
PASS Property vertical-align has initial value baseline
PASS Property vertical-align does not inherit
Harness: the test ran to completion.
